Xact

+++ helper/src/main/java/com/vsg/helper/common/util/viewModel/IViewModelGetDefaultRelation.kt
@@ -2,6 +2,6 @@ package com.vsg.helper.common.util.viewModel
 
 import com.vsg.helper.common.model.IEntity
 
-interface IViewModelGetDefaultRelation<T: IEntity> {
-    fun viewModelGetDefault(idRelation: Long): T?
+interface IViewModelGetDefaultRelation<T : IEntity> {
+    fun viewModelGetDefault(idRelation: Int): T?
 }
